Further separate out structs into own files
