+++ app/models/pokemon.go
@@ -1,4 +1,4 @@
-package app
+package models
 
 type Pokemon struct {
 	Name      string    `njson:"name"`
@@ -7,15 +7,4 @@ type Pokemon struct {
 	Weight    int       `njson:"weight"`
 	Types     []Type    `njson:"types"`
 	Abilities []Ability `njson:"abilities"`
-}
-
-type Type struct {
-	Slot uint8  `njson:"slot"`
-	Name string `njson:"type.name"`
-}
-
-type Ability struct {
-	Slot     uint8  `njson:"slot"`
-	Name     string `njson:"ability.name"`
-	IsHidden bool   `njson:"is_hidden"`
 }
