+++ app/models/type.go
@@ -1,0 +1,6 @@
+package models
+
+type Type struct {
+	Slot uint8  `njson:"slot"`
+	Name string `njson:"type.name"`
+}
